requirements.txt
beautifulsoup4==4.12.2
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
spotipy==2.25.0
